code refactor and ui improvements based on adel feedback
