+++ Assets/Scripts/LevelSelect/ModalController.cs
@@ -4,16 +4,46 @@ using UnityEngine;
 
 public class ModalController : MonoBehaviour
 {
-    // Start is called before the first frame update
+    public GameObject modal;
     public TMPro.TextMeshPro levelText;
     public TMPro.TextMeshPro scoreText;
     public string levelValue;
     public string scoreValue;
-    
+    public GameObject overlay;
+ 
     public void SetValues()
     {
         levelText.text = levelValue;
         scoreText.text = scoreValue;
     }
-    
+    public void ShowLevelModal(GameObject sourceLevel)
+    {
+        //set modal data menjadi level data
+        var levelData = sourceLevel.GetComponent<LevelItem>().data;
+        modal.GetComponent<StarCounter>().StarCount = levelData.starCount;
+        this.scoreValue = levelData.highScore.ToString();
+        this.levelValue = "Level " + levelData.getLevelCount();
+        //TODO: - get level ID then set modal data menjadi level ID
+
+        if (!modal.activeSelf)
+        {
+            modal.GetComponent<StarCounter>().FillStars();
+            this.SetValues();
+            AnimationUtilities.AnimatePopUp(modal);
+            modal.SetActive(true);
+            overlay.SetActive(true);
+        }
+    }
+    public void CloseModal()
+    {
+        AnimationUtilities.AnimatePopUpDisappear(modal);
+        Invoke(nameof(DeactivateModal), 0.2f);  //biar animasinya keplay dulu sebelom diclose
+    }
+    private void DeactivateModal()
+    {
+        modal.GetComponent<StarCounter>().EmptyStars();
+        modal.SetActive(false);
+        overlay.SetActive(false);
+    }
+
 }
